requirements.txt
streamlit
python-docx
langchain
langchain-google-genai
faiss-cpu
fuzzywuzzy[speedup]
python-dotenv
requests
